Move stopAudio() to MediaWidget
